requirements.txt
python-telegram-bot
google-cloud-firestore
google-auth
google-api-core
grpcio
protobuf
requests